REF: made controllers and added styling
