+++ views/pages/login.ejs
@@ -4,18 +4,21 @@
     <header>
         <%- include("../partials/header") %>
     </header>
-    <main>
-        <h1>Hello</h1>
-        <form action="/register/login" method="post">
-            <label for="email">E-mail:</label>
-            <input type="email" id="email" name="email" required>
+    <main class="maintristan">
+        <section class="login">
+            <h1>Login here</h1>
+            <form action="/register/login" method="post" class="loginform">
+                <label for="email">E-mail:</label>
+                <input type="email" id="email" name="email" required>
 
-            <label for="password">Password:</label>
-            <input type="password" id="password" name="password" required>
+                <label for="password">Password:</label>
+                <input type="password" id="password" name="password" required>
 
-            <button type="submit">Login</button>
-        </form>
-        <p>Don't have an account yet?</p><a href="/register">Register</a>
+                <button type="submit">Login</button>
+            </form>
+            <p>Don't have an account yet?</p><a href="/register">Register</a>
+        </section>
     </main>
+    <script src='/scripts/navbarScript.js'></script>
 </body>
 </html>
