+++ views/pages/register.ejs
@@ -4,34 +4,37 @@
     <header>
         <%- include("../partials/header") %>
     </header>
-    <main>
-        <form action="/register" method="post">
-            <label for="firstName">First Name:</label>
-            <input type="text" id="firstName" placeholder="First name" name="firstName" required>
+    <main class="maintristan">
+        <section class="register">
+            <h1>Register your account here</h1>
+            <form action="/register" method="post" class="registerform">
+                <label for="firstName">First Name:</label>
+                <input type="text" id="firstName" placeholder="First name" name="firstName" required>
 
-            <label for="lastName">Last name:</label>
-            <input type="text" id="lastName" placeholder="Last name" name="lastName" required>
+                <label for="lastName">Last name:</label>
+                <input type="text" id="lastName" placeholder="Last name" name="lastName" required>
 
-            <label for="age">Age:</label>
-            <input type="text" id="age" name="age" required>
+                <label for="age">Age:</label>
+                <input type="text" id="age" name="age" placeholder="Age" required>
 
-            <label for="email">E-mail:</label>
-            <input type="email" id="email" name="email" placeholder="example@gmail.com"required>
+                <label for="email">E-mail:</label>
+                <input type="email" id="email" name="email" placeholder="example@gmail.com"required>
 
-            <label for="password">Password:</label>
-            <input type="password" id="password" name="password" placeholder="Password"required>
+                <label for="password">Password:</label>
+                <input type="password" id="password" name="password" placeholder="Password"required>
 
-            <label for="password2">Password check:</label>
-            <input type="password" id="password2" name="password2" placeholder="Re-enter password"required>
+                <label for="password2">Password check:</label>
+                <input type="password" id="password2" name="password2" placeholder="Re-enter password"required>
 
 
 
-            <button type="submit">Register</button>
-        </form>
+                <button type="submit">Register</button>
+            </form>
 
-        <p>Already have an account?</p><a href="register/login">Login</a>
-
+            <p>Already have an account?</p><a href="register/login">Login</a>
+        </section>
     </main>
+    <script src='/scripts/navbarScript.js'></script>
 </body>
 </html>
