+++ views/partials/header.ejs
@@ -4,6 +4,9 @@
         <a class="nav-link" href="/register">Register</a>
       </li>
       <li class="nav-item">
+        <a class="nav-link" href="/register/login">Login</a>
+      </li>
+      <li class="nav-item">
         <a class="nav-link" href="/following/explore">Explore</a>
       </li>
       <li class="nav-item">
